+++ src/main/java/com/aixm/delorean/core/gis/type/components/DistanceType.java
@@ -1,19 +1,21 @@
 package com.aixm.delorean.core.gis.type.components;
 
+import java.math.BigDecimal;
+
 import com.aixm.delorean.core.util.DistanceUom;
 
 import jakarta.persistence.Embeddable;
 
 @Embeddable
 public class DistanceType {
-    protected Double value;
+    protected BigDecimal value;
     protected DistanceUom uom;
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(Double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
